Task02 changed type of vars to double

+++ 02 Classwork/Task02.java
@@ -2,9 +2,9 @@ public class Task02 {
 	
 	public static void main(String[] args) {
 		
-		int a = Integer.parseInt(args[0]);
-		int b = Integer.parseInt(args[1]);
-		int max = a;
+		double a = Double.parseDouble(args[0]);
+		double b = Double.parseDouble(args[1]);
+		double max = a;
 		
 		if (b > max) {
 			max = b;
